parser/tests/tree_construction/basic.dat

#data
<svg><font font=sans></font></svg>
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg font>
|         font="sans"
